requirements.txt
kagglehub
pandas
numpy
pillow
matplotlib
scikit-learn
joblib>=1.2.0
